+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # ts-async-bootstrap
 
+## [2.1.0]
+
+### Additions
+- [Issue #11] onComplete should accept functions that return Promise
+- [Issue #10] BootstrapFunction should accept return type void or Promise<void>
+- [Issue #9] Add bootstrap class for cleaner extension
+- [Issue #8] Add teardown method and handlers
+- [Issue #7] Add onFinally method
+
+### Fixes
+- [Issue #6] Readme example should show import
+
 ## [2.0.0]
 
 ### Breaking Changes